Methoden gefüllt mit MockDaten für Activities
