--- app/src/main/java/com/example/robin/angrynerds_wip/data/services/.gitkeep
@@ -1,2 +1,0 @@
-will contain services that can be called in the activities
-Vorschlag: Trennung zwischen DataServices(get) und ActionServices(set)
