--- app/src/main/java/com/example/robin/angrynerds_wip/data/services/getterService.java
@@ -1,51 +1,0 @@
-package com.example.robin.angrynerds_wip.data.services;
-
-import com.example.robin.angrynerds_wip.data.models.tens.Event;
-import com.example.robin.angrynerds_wip.data.models.tens.Note;
-import com.example.robin.angrynerds_wip.data.models.tens.TEN;
-import com.example.robin.angrynerds_wip.data.models.tens.Todo;
-
-public class getterService {
-
-    /*
-    //TODO tatsächliche Objekte zurückgeben
-    public static TEN[] getAll(){
-        TEN[] arTENs = null;
-        return arTENs;
-    }
-
-    public static TEN getByID(int id){
-        TEN ten = null;
-        return ten;
-    }
-
-    public static Todo getTodoByID(int id) {
-        Todo todo = null;
-        return todo;
-    }
-
-    public static Event getEventByID(int id) {
-        Event event = null;
-        return event;
-    }
-
-    public static Note getNoteByID(int id) {
-        Note note = null;
-        return note;
-    }
-	
-	public static Todo[] getAllTodos(){
-        Todo[] arTodos = null;
-        return arTodos;
-    }
-
-    public static Event[] getAllEvents(){
-        Event[] arEvents = null;
-        return arEvents;
-    }
-
-    public static Note[] getAllNotes(){
-        Note[] arNotes = null;
-        return arNotes;
-    }*/
-}
